move table.rs from interpreter/common to pipeline/common
