+++ src/query/service/src/pipelines/common/table.rs
