+++ src/query/service/src/pipelines/pipeline_builder.rs
@@ -92,8 +92,8 @@ use super::processors::TransformExpandGroupingSets;
 use crate::api::DefaultExchangeInjector;
 use crate::api::ExchangeInjector;
 use crate::interpreters::copy_commit_data_and_set_finish;
-use crate::interpreters::fill_missing_columns;
 use crate::interpreters::PlanParam;
+use crate::pipelines::common::fill_missing_columns;
 use crate::pipelines::processors::transforms::build_partition_bucket;
 use crate::pipelines::processors::transforms::AggregateInjector;
 use crate::pipelines::processors::transforms::FinalSingleStateAggregator;
